refactor: split view tests, rename `job_seeker_infos` include, switch `property` to `cached_property`, clean css & templates
