+++ itou/static/css/itou.css
@@ -399,3 +399,11 @@ Usage:
         width: auto !important;
     }    
 }
+
+.scrolldown-anchor {
+    /* navigation : class to force the scroll down and avoid the banner to hide the map */
+    display: block;
+    position: relative;
+    top: -130px;
+    visibility: hidden;
+}
