+++ itou/templates/siae_evaluations/includes/siae_infos.html
@@ -8,9 +8,7 @@
     </div>
     <div class="col-md-3 text-right">
         <p class="small mb-0">
-            {% if evaluated_siae.state == "PENDING" %}
-                <p class="badge badge-emploi">en attente</b></p>
-            {% endif %}
-        </p>
-    </div>
+            <p class="badge badge-emploi">En attente</b></p>
+    </p>
+</div>
 </div>
