+++ itou/templates/siae_evaluations/institution_evaluated_siae_list.html
@@ -3,20 +3,10 @@
 {% load bootstrap4 %}
 {% load static %}
 
-{% block title %}{% endblock %}
-
-
-{% block messages %}
-    {% bootstrap_messages %}
-{% endblock %}
-
 {% block content %}
-    <div class="row">
+    <div class="row justify-content-center">
 
-        <div class="col-2 ">
-            {# placeholder #}
-        </div>
-        <div class="col-8">
+        <div class="col-md-8" >
             <div class="row mt-3">
                 <div class="col-12">
