+++ itou/templates/siae_evaluations/siae_job_applications_list.html
@@ -12,11 +12,7 @@
 
 {% block content %}
 
-    <div class="row">
-
-        <div class="col-2 ">
-            {# placeholder #}
-        </div>
+    <div class="row justify-content-center">
         <div class="col-8">
 
             <h1 class="h1">
